moved calculator to seperate module, seperated parsing
