+++ index.js
@@ -1,4 +1,3 @@
-
 var path = require('path');
 var data = require('./data');
 
@@ -19,22 +18,4 @@ module.exports.symbols = data.reduce(function(obj, element) {
 	return obj;
 }, {});
 
-module.exports.calculate = {}
-module.exports.calculate.atomicMass = function(formula) {
-  var elements = formula.split(" ");
-  var result = elements.reduce(function(sum, element) {
-    var firstDigit = element.search(/\d/);
-    var factor = 1;
-    var symbol = element;
-    if(firstDigit != -1) {
-      factor = parseInt(element.slice(firstDigit));
-      symbol = element.slice(0, firstDigit);
-    }
-    var e = module.exports.symbols[symbol];
-    var atomicMass = e.atomicMass.split("(")[0];
-    atomicMass = parseFloat(atomicMass);
-    sum += atomicMass * factor;
-    return sum;
-  }, 0);
-  return result;
-};
+module.exports.calculate = require("./calculate");
